pyproject.toml
[tool.poetry]
name = "Online Model Query"
version = "0.1.0"
description = ""
authors = ["Po-han Li <pohanli@utexas.edu>"]

packages = [{include = "query"}]

[tool.poetry.dependencies]
python = "3.9.17"
numpy = "*"
matplotlib = "*"
seaborn = "*"
pandas = "*"
argparse = "*"
black = "*"
scikit-learn = "1.3.0"
protobuf = "3.20.3"
torch = {url="https://download.pytorch.org/whl/cu117/torch-2.0.0%2Bcu117-cp39-cp39-linux_x86_64.whl"}
torchvision = "0.15.1"
transformers = "4.32.0"
datasets = "2.13.1"
nltk = "*"
clip = { git = "https://github.com/openai/CLIP.git"}
pillow = "10.0.0"
stable-baselines3 = "2.0.0"
gymnasium = "0.28.1"
cloudpickle = "2.2.1"
tensorboard = "2.13.0"
tqdm = "4.66.1"
rich = "13.5.2"
contextualbandits = "*"
opencv-python = "*"
evaluate = "0.4.0"
sentencepiece = "^0.1.99"
jupyter = "^1.0.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
black = { version = "*", allow-prereleases = true }
pytest-cov = "*"

# poetry run pip install torch==2.0.0+cu118 torchvision==0.12.0+cu113 -f https://download.pytorch.org/whl/cu118/torch_stable.html
# poetry run pip install --upgrade "jax[cuda]" -f https://storage.googleapis.com/jax-releases/jax_releases.html
# poetry run pip install gym[all]
# poetry run pip install "stable-baselines[mpi]"
# sudo apt install python3.9-dev libglew-dev patchelf
# sudo apt-get install cmake libopenmpi-dev python3-dev zlib1g-dev

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"